remove old router implementation code
