+++ src/router/Route.ts
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Router } from './Router.js';
 
 type Component<P extends {}> = React.FC<P>;
 
@@ -6,35 +7,20 @@ type Component<P extends {}> = React.FC<P>;
  * An immuatable Route structure that can be rendered
  * via react.
  */
-export class Route<P extends {} = any> {
-  private readonly component: Component<P>;
-  private readonly props?: P;
+export class Route<P extends {} = {}> {
+  readonly router: Router;
+  readonly params: P;
+  readonly component: Component<P>;
 
-  constructor(component: Component<P>, props?: P) {
+  constructor(component: Component<P>, params = {} as P) {
     this.component = component;
-    this.props = props;
+    this.params = params;
   }
 
   /**
    * Create react element for rendering
    */
   createElement() {
-    return React.createElement(this.component, this.props);
-  }
-
-  /**
-   * Check if this route identifies as the given Component
-   * @param comp
-   */
-  is(comp: Component<P>) {
-    return this.component === comp;
-  }
-
-   /**
-   * Check if this route identifies as the given Route
-   * @param route
-   */
-  equals(route: Route) {
-    return route && this.component === route.component;
+    return React.createElement(this.component, this.params);
   }
 }
