+++ src/router/index.ts
@@ -1,5 +1,5 @@
-export { Router, defaultRouter } from './Router.js';
-export { Route } from './Route.js';
-
-export * from './hooks.js';
 export { Portal } from './Portal.js';
+export type { Router } from './Router.js';
+export { useRouter, useRoute } from './context.js';
+export { withRouter } from './withRouter.js';
+export { Route } from './Route.js';
