--- src/v2-router/index.ts
@@ -1,3 +1,0 @@
-export { Portal } from './Portal.js';
-export type { Router } from './Router.js';
-export { useRouter, useQueryParams } from './context.js';
